status/functions/pmc_scan_repos/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
python-gnupg
click
python-debian
requests
repoaudit
more-itertools